feat: a simple footer action render
