+++ src/lib/kbar/KBarFooter.tsx
@@ -1,0 +1,152 @@
+import * as React from "react";
+import {Action} from "@/lib/kbar/types";
+import {useEffect, useMemo, useState} from "react";
+import {ActionImpl} from "@/lib/kbar/action";
+import {useKeyPress} from "ahooks";
+import {useKBar} from "@/lib/kbar/useKBar";
+import * as Popover from "@radix-ui/react-popover";
+import RenderItem from "@/lib/kbar/RenderItem";
+
+export interface KBarFooterProps {
+    actions?: (current: Action) => Action[]
+    icon: string | React.ReactElement
+    content: (current: Action) => string | React.ReactElement
+}
+
+const KBarFooter: React.FC<{
+    current?: Action,
+    footer: KBarFooterProps
+    onSubCommandShow: () => void
+    onSubCommandHide: () => void
+}> = ({current, footer, onSubCommandShow, onSubCommandHide}) => {
+    if (!current) {
+        return <>
+        </>
+    }
+
+    const actions = useMemo(() => {
+        return footer.actions ? footer.actions(current).map(a => ActionImpl.create(a, {
+            store: {}
+        })) : []
+    }, [current, footer.actions]);
+
+    return <>
+        <div className='kbar-footer-icon'>
+            {footer.icon}
+        </div>
+        <div className='kbar-footer-content'>
+            {footer.content(current)}
+        </div>
+
+        {
+            actions ? <>
+                    <KBarFooterHr/>
+                    <KBarFooterActions
+                        onSubCommandShow={onSubCommandShow}
+                        onSubCommandHide={onSubCommandHide}
+                        actions={actions}
+                    />
+                </>
+                : <></>
+        }
+    </>
+}
+
+export const KBarFooterHr: React.FC = () => {
+    return (
+        <hr className='kbar-footer-hr'/>
+    );
+}
+
+const KBarFooterActions: React.FC<{
+    actions: ActionImpl[],
+    initialOpen?: boolean,
+    initialShortcut?: string // default 'ctrl.k'
+    onSubCommandShow: () => void
+    onSubCommandHide: () => void
+}> = ({
+          actions,
+          initialOpen,
+          initialShortcut,
+          onSubCommandShow,
+          onSubCommandHide
+      }) => {
+    const [open, setOpen] = React.useState(initialOpen || false)
+    const [shortcut, setShortcut] = React.useState(initialShortcut || 'ctrl.k')
+    const changeVisible = () => setOpen((o) => !o)
+    useKeyPress(shortcut, (e) => {
+        e.preventDefault()
+        changeVisible()
+    })
+    React.useEffect(() => {
+        if (open) {
+            query.getInput().blur()
+            onSubCommandShow()
+        }
+    }, [open])
+
+    const {query} = useKBar()
+
+    return <Popover.Root open={open} onOpenChange={setOpen} modal>
+        <Popover.Trigger className='kbar-subcommand-trigger' onClick={() => setOpen(true)} aria-expanded={open}>
+            <span>Actions</span>
+            {shortcut.split('.').map((s, i) => <kbd key={i}>{s}</kbd>)}
+        </Popover.Trigger>
+        <Popover.Content
+            side="top"
+            align="end"
+            className="kbar-subcommand-menu"
+            sideOffset={16}
+            alignOffset={0}
+            onCloseAutoFocus={(e) => {
+                e.preventDefault()
+                query.getInput().focus()
+                onSubCommandHide()
+            }}
+        >
+            <div className='kbar-subcommand-menu-content'>
+                <RenderFooterActions actions={actions}/>
+            </div>
+        </Popover.Content>
+    </Popover.Root>
+}
+
+
+const RenderFooterActions = ({actions}: { actions: ActionImpl[] }) => {
+    const [selectedActionIndex, setSelectedActionIndex] = useState(0);
+
+    const handleKeyDown = (e: KeyboardEvent) => {
+        if (e.key === 'ArrowUp') {
+            setSelectedActionIndex(prevIndex => (prevIndex > 0 ? prevIndex - 1 : actions.length - 1));
+        } else if (e.key === 'ArrowDown') {
+            setSelectedActionIndex(prevIndex => (prevIndex < actions.length - 1 ? prevIndex + 1 : 0));
+        } else if (e.key === 'Enter') {
+            const selectedAction = actions[selectedActionIndex];
+            if (selectedAction) {
+                console.log("run it ", selectedAction)
+            }
+        }
+    };
+
+    useEffect(() => {
+        window.addEventListener('keydown', handleKeyDown);
+        return () => {
+            window.removeEventListener('keydown', handleKeyDown);
+        };
+    }, [selectedActionIndex]);
+
+    return (
+        <ul>
+            {actions.map((action, index) => (
+                <RenderItem
+                    active={index === selectedActionIndex}
+                    action={action}
+                    currentRootActionId={''}
+                />
+            ))}
+        </ul>
+    );
+};
+
+
+export default KBarFooter
